Fix big flaw regarding creating many same-ish names in a folder
Basically, if you had a lot of files that started with "longfilename_*.txt" (quite common in game files) it would basically have to do a very long process of reiterating what previous tails are already used. If you have 200 files, and it has to do 200 file checks, which is super slow. This dumb algorithm is something that windows used to use in the old days, but which is now replaced with a checksum approach which basically guarantees that you never have to reiterate the file checksums.

https://en.wikipedia.org/wiki/8.3_filename#VFAT_and_Computer-generated_8.3_filenames is an interesting page to read about this alias and numeric tail. The previous thing was like the method 5 that's mentioned, while now it uses method 4.